+++ components/camel-braintree/README.md
@@ -18,10 +18,25 @@ privateKey  = private key from sandbox ui
 Then, you can use the following commands to run the tests:
 
 ```
-CAMEL_BRAINTREE_REPORT_DATE=$(date '+%Y-%m-%d') CAMEL_BRAINTREE_MERCHANT_ACCOUNT_ID="merchant ID taken from sandbox" ui mvn -Pbraintree-test clean verify
+CAMEL_BRAINTREE_REPORT_DATE=$(date '+%Y-%m-%d') CAMEL_BRAINTREE_MERCHANT_ACCOUNT_ID="merchant ID taken from sandbox" ui mvn mvn -DbraintreeAuthenticationType=PUBLIC_PRIVATE_KEYS clean verify 
 ```
 
 It's also possible to run a smaller set of tests by running them without the environment variables:
 ```
-mvn -Pbraintree-test clean verify
+mvn -DbraintreeAuthenticationType=PUBLIC_PRIVATE_KEYS clean verify
+```
+
+# Authentication types
+
+The authentication type can be passed via `braintreeAuthenticationType` system property. 
+You can use one of: 
+
+* PUBLIC_PRIVATE_KEYS
+* ACCESS_TOKEN
+
+
+I.e.:
+
+```
+mvn -DbraintreeAuthenticationType=ACCESS_TOKEN clean verify
 ```
