+++ components/camel-braintree/src/test/java/org/apache/camel/component/braintree/integration/AddOnGatewayIT.java
@@ -14,29 +14,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.braintree;
+package org.apache.camel.component.braintree.integration;
 
 import java.util.List;
 
-import com.braintreegateway.Plan;
+import com.braintreegateway.AddOn;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.braintree.internal.PlanGatewayApiMethod;
+import org.apache.camel.component.braintree.AbstractBraintreeTestSupport;
+import org.apache.camel.component.braintree.internal.AddOnGatewayApiMethod;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class PlanGatewayIntegrationTest extends AbstractBraintreeTestSupport {
+@EnabledIfSystemProperty(named = "braintreeAuthenticationType", matches = ".*")
+public class AddOnGatewayIT extends AbstractBraintreeTestSupport {
 
-    private static final Logger LOG = LoggerFactory.getLogger(PlanGatewayIntegrationTest.class);
-    private static final String PATH_PREFIX = getApiNameAsString(PlanGatewayApiMethod.class);
+    private static final Logger LOG = LoggerFactory.getLogger(AddOnGatewayIT.class);
+    private static final String PATH_PREFIX = getApiNameAsString(AddOnGatewayApiMethod.class);
 
     @Disabled
     @Test
     public void testAll() throws Exception {
-        final List<Plan> result = requestBody("direct://ALL", null, List.class);
+        final List<AddOn> result = requestBody("direct://ALL", null, List.class);
 
         assertNotNull(result, "all result");
         LOG.debug("all: " + result);
