+++ components/camel-braintree/src/test/resources/test-options.properties
@@ -16,8 +16,8 @@
 ## ---------------------------------------------------------------------------
 
 #environment = SANDBOX
-#merchantId  = 
-#publicKey   = 
+#merchantId  =
+#publicKey   =
 #privateKey  =
 #accessToken =
